Fixed some type issues with sqrt. Fixed rect.isValid to work with all types and avoid useless tests. Cleaned up some variable uses and scopes.
git-svn-id: svn://svn.code.sf.net/p/irrlicht/code/trunk@1004 dfc29bdd-3216-0410-991c-e03cc46cb475

+++ include/quaternion.h
@@ -472,7 +472,7 @@ inline void quaternion::fromAngleAxis(f32 angle, const vector3df& axis)
 
 inline void quaternion::toAngleAxis(f32 &angle, core::vector3df &axis) const
 {
-	f32 scale = sqrt (X*X + Y*Y + Z*Z);
+	f32 scale = sqrtf(X*X + Y*Y + Z*Z);
 
 	if (core::iszero(scale) || W > 1.0f || W < -1.0f)
 	{
